Protocol/Fetch.Ai/requirements.txt
uagents==0.22.8
Flask
requests
ipfshttpclient
mnemonic
openai>=1.0.0
flask-cors
numpy
paho-mqtt
rapidfuzz
aiohttp
web3
python-dotenv

